The settings are separated into its own file
